+++ tests/phpunit/Components/Validator/MixedStructureValidatorTest.php
@@ -6,6 +6,7 @@ use PHPUnit\Framework\TestCase;
 use PHPUnuhi\Bundles\Storage\JSON\JsonStorage;
 use PHPUnuhi\Components\Validator\MissingStructureValidator;
 use PHPUnuhi\Models\Configuration\Filter;
+use PHPUnuhi\Models\Configuration\Protection;
 use PHPUnuhi\Models\Translation\Locale;
 use PHPUnuhi\Models\Translation\TranslationSet;
 
@@ -92,6 +93,7 @@ class MixedStructureValidatorTest extends TestCase
         return new TranslationSet(
             '',
             'json',
+            new Protection(),
             $locales,
             new Filter(),
             [],