+++ tests/phpunit/Components/Validator/Rules/DisallowedTextsRuleTest.php
@@ -6,6 +6,7 @@ use PHPUnit\Framework\TestCase;
 use PHPUnuhi\Bundles\Storage\JSON\JsonStorage;
 use PHPUnuhi\Components\Validator\Rules\DisallowedTextsRule;
 use PHPUnuhi\Models\Configuration\Filter;
+use PHPUnuhi\Models\Configuration\Protection;
 use PHPUnuhi\Models\Translation\Locale;
 use PHPUnuhi\Models\Translation\TranslationSet;
 
@@ -87,6 +88,7 @@ class DisallowedTextsRuleTest extends TestCase
         return new TranslationSet(
             '',
             'json',
+            new Protection(),
             $locales,
             new Filter(),
             [],