+++ tests/phpunit/Components/Validator/Rules/DuplicateContentRuleTest.php
@@ -6,6 +6,7 @@ use PHPUnit\Framework\TestCase;
 use PHPUnuhi\Bundles\Storage\JSON\JsonStorage;
 use PHPUnuhi\Components\Validator\Rules\DuplicateContentRule;
 use PHPUnuhi\Models\Configuration\Filter;
+use PHPUnuhi\Models\Configuration\Protection;
 use PHPUnuhi\Models\Translation\Locale;
 use PHPUnuhi\Models\Translation\TranslationSet;
 
@@ -104,6 +105,7 @@ class DuplicateContentRuleTest extends TestCase
         return new TranslationSet(
             '',
             'json',
+            new Protection(),
             $locales,
             new Filter(),
             [],