+++ tests/phpunit/Components/Validator/Rules/MaxKeyLengthRuleTest.php
@@ -6,6 +6,7 @@ use PHPUnit\Framework\TestCase;
 use PHPUnuhi\Bundles\Storage\JSON\JsonStorage;
 use PHPUnuhi\Components\Validator\Rules\MaxKeyLengthRule;
 use PHPUnuhi\Models\Configuration\Filter;
+use PHPUnuhi\Models\Configuration\Protection;
 use PHPUnuhi\Models\Translation\Locale;
 use PHPUnuhi\Models\Translation\TranslationSet;
 
@@ -103,7 +104,6 @@ class MaxKeyLengthRuleTest extends TestCase
 
     /**
      * @param array $locales
-     * @param int $keyLength
      * @return TranslationSet
      */
     private function buildSet(array $locales): TranslationSet
@@ -111,6 +111,7 @@ class MaxKeyLengthRuleTest extends TestCase
         return new TranslationSet(
             '',
             'json',
+            new Protection(),
             $locales,
             new Filter(),
             [],