+++ tests/phpunit/Components/Validator/Rules/NestingDepthRuleTest.php
@@ -6,6 +6,7 @@ use PHPUnit\Framework\TestCase;
 use PHPUnuhi\Bundles\Storage\JSON\JsonStorage;
 use PHPUnuhi\Components\Validator\Rules\NestingDepthRule;
 use PHPUnuhi\Models\Configuration\Filter;
+use PHPUnuhi\Models\Configuration\Protection;
 use PHPUnuhi\Models\Configuration\Rule;
 use PHPUnuhi\Models\Configuration\Rules;
 use PHPUnuhi\Models\Translation\Locale;
@@ -103,6 +104,7 @@ class NestingDepthRuleTest extends TestCase
         return new TranslationSet(
             '',
             'json',
+            new Protection(),
             $locales,
             new Filter(),
             [],