+++ tests/phpunit/Models/Configuration/ConfigurationTest.php
@@ -5,6 +5,7 @@ namespace phpunit\Models\Configuration;
 use PHPUnit\Framework\TestCase;
 use PHPUnuhi\Models\Configuration\Configuration;
 use PHPUnuhi\Models\Configuration\Filter;
+use PHPUnuhi\Models\Configuration\Protection;
 use PHPUnuhi\Models\Translation\TranslationSet;
 
 class ConfigurationTest extends TestCase
@@ -16,7 +17,7 @@ class ConfigurationTest extends TestCase
     public function testTranslationSets()
     {
         $sets = [];
-        $sets[] = new TranslationSet('products', 'ini', [], new Filter(), [], [], []);
+        $sets[] = new TranslationSet('products', 'ini', new Protection(), [], new Filter(), [], [], []);
 
         $config = new Configuration($sets);